Update snapshot with launch tempalte

+++ infra/deploy.sh
@@ -11,9 +11,13 @@ if [ -z "${NEW_AMI}" ]; then
 	echo "[!] NEW_AMI is not set."
 	exit
 fi
+if [ -z "${NEW_SNAPSHOT_ID}" ]; then
+	echo "[!] NEW_SNAPSHOT_ID is not set."
+	exit
+fi
 
 # create new version
-aws ec2 create-launch-template-version --launch-template-name "${LAUNCH_TEMPLATE_NAME}" --source-version '$Latest' --launch-template-data '{"ImageId":"'"${NEW_AMI}"'"}' > /dev/null
+aws ec2 create-launch-template-version --launch-template-name "${LAUNCH_TEMPLATE_NAME}" --source-version '$Latest' --launch-template-data '{"ImageId":"'"${NEW_AMI}"'", "BlockDeviceMappings": [{"Ebs": {"SnapshotId": "'"${NEW_SNAPSHOT_ID}"'"}}]}' > /dev/null
 
 # delete policy
 aws autoscaling delete-policy --auto-scaling-group-name "${SCALING_GROUP_NAME}" --policy-name "${SCALING_POLICY_NAME}"
@@ -36,7 +40,7 @@ if [ "$current_instances" -ne "$desired_instances" ] || [ "$desired_instances" -
 fi
 
 # scale up
-desired_instances_transition_phase="$(expr "$desired_instances" '*' 2)"
+desired_instances_transition_phase="$((desired_instances * 2))"
 aws autoscaling update-auto-scaling-group --auto-scaling-group-name shipvote-web --default-cooldown 20 > /dev/null
 aws autoscaling set-desired-capacity --auto-scaling-group-name "${SCALING_GROUP_NAME}" --desired-capacity "$desired_instances_transition_phase" > /dev/null
 
@@ -64,7 +68,7 @@ while true; do
 	    aws autoscaling put-scaling-policy --auto-scaling-group-name "${SCALING_GROUP_NAME}" --policy-name "${SCALING_POLICY_NAME}" --cli-input-json '{"PolicyType":"TargetTrackingScaling", "TargetTrackingConfiguration":{"TargetValue":85, "DisableScaleIn":false, "PredefinedMetricSpecification":{"PredefinedMetricType":"ASGAverageCPUUtilization"}}, "EstimatedInstanceWarmup":300}'
         exit 1
     fi
-    let tries=tries+1
+    ((tries = tries + 1))
 
     healthy_instances="$(aws elbv2 describe-target-health --target-group-arn "${target_group_arn}" | jq '[.TargetHealthDescriptions[] | select(.TargetHealth.State == "healthy")] | length')"
